+++ react/src/hooks/hooks.ts
@@ -1,11 +1,6 @@
-import { useState, useCallback, useEffect } from "react";
+import { useState, useCallback } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import {
-  useToast,
-  UseToastOptions,
-  ToastPositionWithLogical,
-  useMediaQuery,
-} from "@chakra-ui/react";
+import { UseToastOptions, ToastPositionWithLogical } from "@chakra-ui/react";
 
 export interface INotificationProps extends UseToastOptions {
   title?: string;
@@ -13,32 +8,6 @@ export interface INotificationProps extends UseToastOptions {
   type?: "info" | "warning" | "success" | "error" | undefined;
   position?: ToastPositionWithLogical;
 }
-
-export const useNotification = () => {
-  const toast = useToast();
-  const [isSmallScreen] = useMediaQuery("(max-width: 1070px)");
-  const [position, setPosition] = useState<ToastPositionWithLogical>("bottom");
-
-  useEffect(() => {
-    setPosition(isSmallScreen ? "bottom" : "top");
-  }, [isSmallScreen]);
-
-  const createToast = useCallback(
-    ({ message, type, ...props }: INotificationProps) =>
-      toast({
-        duration: 5000,
-        isClosable: true,
-        position,
-        variant: "subtle",
-        status: type,
-        description: message,
-        ...props,
-      }),
-    [toast, position],
-  );
-
-  return { createToast };
-};
 
 export const useGetUrlForResourceHelpers = () => {
   const { baseId } = useParams<{ baseId: string }>();
